pyproject.toml
[project]
name = "portfolio-dashboard"
version = "0.1.0"
description = "법인 투자 포트폴리오 대시보드 (모니터링 보조 도구)"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.36",
    "pandas>=2.2",
    "plotly>=5.22",
    "pyyaml>=6.0",
    "yfinance>=0.2.40",
    "pykrx>=1.0.45",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
